requirements.txt
streamlit
azure-functions
azure-functions-durable